+++ changelog.md
@@ -1,6 +1,7 @@
 ** Version 0.5pre - Date 02.05.2011 **
 
   * Fixed a bug in M.I18N
+  * Added setValue() to M.LabelView
 
 ** Version 0.4 - Date 19.04.2011 **